# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Support Desk API",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/sign-in": {
      "post": {
        "summary": "Sign in with an email and a password; answers with a bearer token and the account's public fields.",
        "responses": {
          "200": {
            "description": "Signed in. Body: { token, user: { id, role, name } }."
          },
          "400": {
            "description": "The body was not JSON."
          },
          "401": {
            "description": "Wrong password, unknown address, or a soft-deleted account — the body does not distinguish them."
          },
          "422": {
            "description": "Well-formed JSON, but the credentials were the wrong shape. The body names the fields and never their values."
          },
          "429": {
            "description": "Too many failed sign-ins for this account or from this network address. The body does not say which, because which one it was would say whether the account exists."
          },
          "500": {
            "description": "Internal."
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "summary": "Whoever the bearer token resolves to. Guarded.",
        "responses": {
          "200": {
            "description": "Body: { id, role, name }."
          },
          "401": {
            "description": "No token, or one that is malformed, tampered with, expired, foreign-signed, or issued to an account since removed."
          },
          "500": {
            "description": "Internal."
          }
        }
      }
    },
    "/api/v1/me/password": {
      "post": {
        "summary": "Change your own password, knowing the current one. One route for every role: an admin doing this uses it too, which is why the admin route for setting somebody else's password refuses a caller their own account — an admin who could skip the current-password check on themselves would turn a stolen session into a permanent one. Body: { currentPassword, newPassword }. Tokens issued before the change stay valid until they expire; ending them is IDENTITY-8-API, and that is a stated gap. Changing a password ends every OTHER session (IDENTITY-8-API): the account records when its password last changed, and the resolver refuses any token issued before that. The token that made THIS request was issued under the old password and is refused too, so the answer carries a fresh one — which is what keeps the session that made the change alive without a rule to tell it apart from the ones being ended. On a whole-second clock such a rule could not be written: a token minted in the same second as the change is accepted, and that window is stated rather than hidden.",
        "responses": {
          "200": {
            "description": "{ id, updatedAt, token }. The token replaces the session this change just ended — a client that ignores it is signed out on its next request. No password and no hash, here or in the audit row."
          },
          "401": {
            "description": "The current password is wrong. Sign-in gives one refusal for three causes so the response cannot be used to learn which addresses exist; that reasoning does not reach here, because this caller is already authenticated and already knows the account exists. Telling them which half they got wrong leaks nothing."
          },
          "422": {
            "description": "`fields` names it: `currentPassword` when it is missing, `newPassword` when it is missing, below the twelve-character floor, or equal to the current one — a change that changes nothing is a change somebody believes they made."
          },
          "500": {
            "description": "Unhandled, logged with the request id."
          }
        }
      }
    },
...
    "/api/v1/accounts": {
      "post": {
        "summary": "Create a staff account. Answers with the account and, once, the password it was given.",
        "responses": {
          "201": {
            "description": "Created. Body: { user, initialPassword } — the password is shown here and never again."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "The subject is not an admin. The service is never entered."
          },
          "500": {
            "description": "Internal."
          },
          "409": {
            "description": "EMAIL_TAKEN — the address already belongs to an account, live or disabled. A domain code rather than a bare CONFLICT: this route and the others below refuse for different reasons, and a screen has to tell them apart."
          },
          "422": {
            "description": "The body was the wrong shape. It names the fields and never their values."
          }
        }
      },
      "get": {
        "summary": "The staff accounts, paginated. `state` chooses which: `live` (the default, and what this route answered before the parameter existed), `disabled`, or `all`. It takes one because nothing here listed a disabled account, so POST /accounts/{id}/re-enable took an id no client had any way to learn — the route was served, documented, tested, and callable by nothing. An unrecognised value is refused rather than treated as one of the three: a listing that quietly answered about everybody when it was asked about the disabled would be worse than one that refuses.",
        "responses": {
          "200": {
            "description": "Body: { items, total, limit, offset, state }. `total` counts what the requested `state` matches, not what is live — a disabled listing beside the live total is a pager wrong on every page but the first. `state` is echoed so a reader of the response knows which question it answers. Every item carries `deletedAt`: null while the account is live, and the moment it was disabled otherwise."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "The subject is not an admin. The service is never entered."
          },
          "500": {
            "description": "Internal."
          },
          "422": {
            "description": "A limit above the ceiling, a malformed window, or a `state` that is not live, disabled or all. Named in `fields`. Refused, not clamped."
          }
        }
      }
    },
    "/api/v1/accounts/{id}/role": {
      "patch": {
        "summary": "Set an account's role.",
        "responses": {
          "200": {
            "description": "Body: { user, changed }. changed is false when the role was already that."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "The subject is not an admin. The service is never entered."
          },
          "500": {
            "description": "Internal."
          },
          "404": {
            "description": "No live account with that id."
          },
          "409": {
            "description": "LAST_ADMIN — the change would leave the desk with nobody who can administer it. Nothing is refused for a role that is simply unchanged; that answers 200 with `changed: false`."
          },
          "422": {
            "description": "The role is not one this system has."
          }
        }
      }
    },
    "/api/v1/accounts/{id}/disable": {
      "post": {
        "summary": "Disable an account. Nothing is deleted; the row keeps its history. Disabling also hands back the tickets that account still has to work on — everything assigned to them that is not closed, unassigned in the SAME transaction as the disable. An account disabled with its queue still assigned to it is worse than either outcome alone: the work is invisible in every unassigned view and its owner cannot sign in to hand it over. A resolved ticket is included because a reply within its window reopens it, and then it is somebody's again; a closed one is not, because rewriting who finished it to tidy a queue would make the record wrong. Each move is audited as an ordinary assignment with the disabling admin as the actor, and each ticket's revision is bumped — so an agent who had one open is refused rather than allowed to write over it.",
        "responses": {
          "200": {
            "description": "{ user, unassigned }. `unassigned` is how many tickets were handed back — zero is an answer, not an omission. An admin deciding whether to disable somebody is deciding what happens to their work, and a number they have to go and count is a number they will not count."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "The subject is not an admin. The service is never entered."
          },
          "500": {
            "description": "Internal."
          },
          "404": {
            "description": "No account with that id."
          },
          "409": {
            "description": "ALREADY_DISABLED when the account is already disabled — two admins pressing the same button, which is not a refusal so much as news. LAST_ADMIN when it would leave the desk with nobody who can administer it."
          }
        }
      }
    },
    "/api/v1/accounts/{id}/re-enable": {
      "post": {
        "summary": "Re-enable a disabled account. It is the same row it always was.",
        "responses": {
          "200": {
            "description": "Body: { user }, with deletedAt cleared."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "The subject is not an admin. The service is never entered."
          },
          "500": {
            "description": "Internal."
          },
          "404": {
            "description": "No account with that id."
          },
          "409": {
            "description": "ALREADY_LIVE when the account was never disabled. EMAIL_TAKEN when a live account has taken the address since — the partial unique index frees it while a row is disabled, so the collision is real and the row stays disabled."
          }
        }
      }
    },
    "/api/v1/accounts/{id}/set-password": {
      "post": {
        "summary": "An admin sets another user's password, so somebody locked out gets back in. There is no reset by email in this product — the brief puts it under \"Specified only\" — which is why a person who forgets theirs needs a colleague rather than a link. The body is { password }; twelve characters is the floor and there are no composition rules, because a rule demanding a symbol produces the same password with a symbol on the end. Nothing comes back but the id and the time: the admin typed it, so there is nothing to read. Tokens issued before this stay valid until they expire — ending them is IDENTITY-8-API, and that is a stated gap rather than an oversight.",
        "responses": {
          "200": {
            "description": "{ id, updatedAt }. No password and no hash, here or in the audit row."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "403": {
            "description": "Not an admin — or an admin naming themselves. Setting somebody else's password is for a person who is locked out; changing your own is IDENTITY-7-API and asks for the current one, and an admin who could skip that check on themselves would turn a stolen session into a permanent one."
          },
          "404": {
            "description": "No user with that id."
          },
          "409": {
            "description": "ALREADY_DISABLED — the account is disabled. Bringing somebody back is re-enable, with its own route and its own audit row."
          },
          "422": {
            "description": "The password is missing or shorter than the floor; `fields` names it."
          },
          "500": {
            "description": "Unhandled, logged with the request id."
          }
        }
      }
    },
    "/api/v1/assignees": {
      "get": {
        "summary": "The live staff a ticket can be assigned to. Any signed-in staff member may read it — an agent who cannot see the list cannot hand a ticket over.",
        "responses": {
          "200": {
            "description": "{ items, total, limit, offset }. Each item is { id, name, role } — no address, no timestamps, nothing about the password."
          },
          "401": {
            "description": "No token, or one that does not resolve to a live account."
          },
          "422": {
            "description": "A limit above the ceiling, or a malformed window. Refused, never clamped (BR-4)."
          },
          "500": {
            "description": "Internal."
          }
        }
      }
    },
...
  }
}